+++ packages/zp/src/index.ts
@@ -1,3 +1,4 @@
 export * from './app'
 export * from './errors'
 export * from './declaration'
+export * from './resolver'