+++ playground/hello/src/mock.ts
@@ -1,2 +1,2 @@
 
-export const people = [{ id: 1, name: 'Alex' }, { id: 2, name: 'Ion' }, { id: 3, name: 'Vasile' }]
+export const people = [{ id: 1, name: 'Alex', age: 37 }, { id: 2, name: 'Ion', age: 10 }, { id: 3, name: 'Vasile', age: 22 }]