+++ packages/zp-data/CHANGELOG.md
@@ -1,5 +1,12 @@
 # zapnode-data
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - zapnode@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes