+++ packages/zp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # zapnode
 
+## 0.0.10
+
+### Patch Changes
+
+- add resolve, resultHook, dataHook
+
 ## 0.0.9
 
 ### Patch Changes