+++ playground/hello/CHANGELOG.md
@@ -1,5 +1,14 @@
 # hello
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - zapnode-plugins@0.0.11
+  - zapnode@0.0.10
+  - zapnode-data@0.0.10
+
 ## 1.0.2
 
 ### Patch Changes